[SR-PY] cleaner mock handle
Summary

Need to do this to mock openai because key doesnt work today
